+++ api/v1alpha1/userconfig/service/opensearch/opensearch.go
@@ -53,20 +53,6 @@ type AzureMigration struct {
 	// The snapshot name to restore from
 	SnapshotName string `groups:"create,update" json:"snapshot_name"`
 }
-type CustomKeystores struct {
-	// +kubebuilder:validation:Pattern=`^[^\r\n]*$`
-	Name string `groups:"create,update" json:"name"`
-
-	// +kubebuilder:validation:Enum="azure";"gcs";"s3"
-	Type string `groups:"create,update" json:"type"`
-}
-type CustomRepos struct {
-	// +kubebuilder:validation:Pattern=`^[^\r\n]*$`
-	Name string `groups:"create,update" json:"name"`
-
-	// +kubebuilder:validation:Enum="azure";"gcs";"s3"
-	Type string `groups:"create,update" json:"type"`
-}
 
 // Google Cloud Storage migration settings
 type GcsMigration struct {
@@ -1078,14 +1064,6 @@ type OpensearchUserConfig struct {
 	// +kubebuilder:validation:MaxLength=255
 	// Serve the web frontend using a custom CNAME pointing to the Aiven DNS name. When you set a custom domain for a service deployed in a VPC, the service certificate is only created for the public-* hostname and the custom domain.
 	CustomDomain *string `groups:"create,update" json:"custom_domain,omitempty"`
-
-	// +kubebuilder:validation:MaxItems=10
-	// Allow to register custom keystores in OpenSearch
-	CustomKeystores []*CustomKeystores `groups:"create,update" json:"custom_keystores,omitempty"`
-
-	// +kubebuilder:validation:MaxItems=10
-	// Allow to register object storage repositories in OpenSearch
-	CustomRepos []*CustomRepos `groups:"create,update" json:"custom_repos,omitempty"`
 
 	// Disable automatic replication factor adjustment for multi-node services. By default, Aiven ensures all indexes are replicated at least to two nodes. Note: Due to potential data loss in case of losing a service node, this setting can not be activated unless specifically allowed for the project.
 	DisableReplicationFactorAdjustment *bool `groups:"create,update" json:"disable_replication_factor_adjustment,omitempty"`